requirements.txt
pyaudio
openai-whisper
torch
ffmpeg-python
tiktoken
numba
setuptools-rust
openai
gtts
playsound
pyttsx3
